feat: 비밀번호 찾기 구현

+++ src/main/java/univcapstone/employmentsite/controller/MyPageController.java
@@ -62,8 +62,6 @@ public class MyPageController {
     ) {
         User user = customUserDetails.getUser();
 
-        log.info("유저 마이페이지");
-
         DefaultResponse<User> defaultResponse = DefaultResponse.<User>builder()
                 .code(HttpStatus.OK.value())
                 .httpStatus(HttpStatus.OK)
@@ -74,16 +72,18 @@ public class MyPageController {
         return ResponseEntity.ok().body(defaultResponse);
     }
 
-    @GetMapping("/user/picture")
-    public ResponseEntity<? extends BasicResponse> myPicture(
+    //프로필 사진만 가져오기(URL)
+    @GetMapping("user/image/show")
+    public ResponseEntity<? extends BasicResponse> getMyProfile(
             @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
+    ){
         User user = customUserDetails.getUser();
-        Map<String,String> image=pictureService.getProfileImageName(user);
-        DefaultResponse<Map<String,String>> defaultResponse = DefaultResponse.<Map<String,String>>builder()
+        String image=pictureService.getProfileImageOne(user);
+
+        DefaultResponse<String> defaultResponse = DefaultResponse.<String>builder()
                 .code(HttpStatus.OK.value())
                 .httpStatus(HttpStatus.OK)
-                .message("아마존에서 온 프로필")
+                .message("사용자의 프로필 사진")
                 .result(image)
                 .build();
 
@@ -91,6 +91,7 @@ public class MyPageController {
                 .body(defaultResponse);
     }
 
+    //프로필 사진 저장
     @PostMapping(value = "/user/image/save", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
     public ResponseEntity<? extends BasicResponse> saveProfile(
             @AuthenticationPrincipal CustomUserDetails customUserDetails,
@@ -116,24 +117,70 @@ public class MyPageController {
         }
 
     }
-    @DeleteMapping(value ="/user/image/delete",consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
+
+    //현재 프로필 사진 삭제
+    @DeleteMapping(value ="/user/image/delete")
     public ResponseEntity<? extends BasicResponse> deleteProfile(
             @AuthenticationPrincipal CustomUserDetails customUserDetails
     ){
-        Picture picture=pictureService.getProfilePicture(customUserDetails.getUser());
-        log.info("filePath= {} ",picture.getUploadFileName());
-        String result=pictureService.deleteFile(picture);
+        List<Picture> pictures=pictureService.getAllProfileImageName(customUserDetails.getUser());
+        for(Picture picture : pictures){
+            if(picture.isProfile()){
+                pictureService.deleteFile(picture);
+            }
+        }
+
         return ResponseEntity.ok(
                 DefaultResponse.builder()
                         .code(HttpStatus.OK.value())
                         .httpStatus(HttpStatus.OK)
                         .message("사진 삭제 완료")
-                        .result(result)
+                        .result("")
                         .build()
         );
     }
 
+    //프로필 사진 업데이트
+    @PatchMapping("/user/image/update")
+    public ResponseEntity<? extends BasicResponse> updateProfile(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails,
+            @RequestPart(value = "files") MultipartFile multipartFiles
+    ) throws IOException {
+        User user = customUserDetails.getUser();
+        List<Picture> pictures=pictureService.getAllProfileImageName(user);
+        for(Picture picture : pictures){
+            if(picture.isProfile()){
+                pictureService.deleteFile(picture);
+            }
+        }
+        String uploadImagesUrl = pictureService.uploadProfileFile(multipartFiles, dirName,user);
 
+        return ResponseEntity.ok(
+                DefaultResponse.builder()
+                        .code(HttpStatus.OK.value())
+                        .httpStatus(HttpStatus.OK)
+                        .message("사진 업데이트 완료")
+                        .result(uploadImagesUrl)
+                        .build()
+        );
+    }
+    //합성(만들어진) 사진 모두를 가져오기
+    @GetMapping("/user/picture")
+    public ResponseEntity<? extends BasicResponse> myPicture(
+            @AuthenticationPrincipal CustomUserDetails customUserDetails
+    ) {
+        User user = customUserDetails.getUser();
+        List<Map<String,String>> image=pictureService.getConversionImage(user);
+        DefaultResponse<List<Map<String,String>>> defaultResponse = DefaultResponse.<List<Map<String,String>>>builder()
+                .code(HttpStatus.OK.value())
+                .httpStatus(HttpStatus.OK)
+                .message("아마존에서 온 합성된 사진들")
+                .result(image)
+                .build();
+
+        return ResponseEntity.ok()
+                .body(defaultResponse);
+    }
     @GetMapping("/user/bookmark")
     public ResponseEntity<? extends BasicResponse> myBookmark(
             @AuthenticationPrincipal CustomUserDetails customUserDetails
